# Introduce a new Performance Monitor and Developer Mode