**NOTE:** It'll be easiest to view this PR with [whitespace ignored](https://github.blog/2018-05-01-ignore-white-space-in-code-review/). ([quick link](https://github.com/pfalstad/circuitjs1/pull/20/files?diff=unified&w=1))

This PR introduces a new `PerformanceMonitor` that uses a high precision timer to get a better view of how the sim is performing. The monitor can be split into different contexts that can be nested.

![2022-04-14 22_37_24-127 0 0 1_8888_circuitjs html](https://user-images.githubusercontent.com/6798456/163527420-c90f536a-e04f-4705-a148-e4a10c656a2d.png)

```
updateCircuit(): 2.99   // The whole sim took 2.299 ms
-runCircuit(): 0.2      // This function took 0.2 ms
-graphics: 2            // This is showing the total of the following two child times (2ms)
--elm.draw(): 0.2       // Took 0.2ms to draw all the elements
--drawBottomArea(): 1.7 // Took 1.7ms to draw the scopes
```

Additionally, this developer info can now be togged without recompiling via a new option in the settings.

![2022-04-14 22_43_19-127 0 0 1_8888_circuitjs html](https://user-images.githubusercontent.com/6798456/163527850-5bb0cec1-8505-4623-9d62-7699289158b4.png)

This PR is very much in line with the kind of contributions I want to make to the project, refactoring + new small features where it makes sense.

@pfalstad would you consider opening the issues page for this fork of the project? Using the sharpie7 issues page dosen't seem appropriate.