@pfalstad 2022-04-20
Sorry, I forgot about this.  Thanks, this would have been handy when I was working on my pong project!  I wasn't sure if I wanted a checkbox in the options, but that is really a small part of it and not a big deal.  I have merged this into my local sandbox and it will show up next time I release.