@Mervill 2022-04-23
I'm working on expanding this, so merging via github will let me prepare new pull requests without having to worry about rebase problems further down the line. Let me know if you have any concerns about merge conflicts with your local changes.

Regarding the Developer Mode option: Commented out code is impossible to maintain, so if you'd prefer not to have an option in the menus, simply having a `developerMode` Boolean in the main class would be sufficient. Eventually I hope developer mode will expose a bunch more info that will make characterizing changes to the internals much easier, so having it all behind a coinvent switch is essential.