@Mervill 2022-04-23
@pfalstad Hey thanks! I'm looking forward to contributing more to the project. The last thing I want to do is step on your toes or overwhelm you with stuff (we're both doing this on the side, after all) so don't hesitate to tell me if I need to cool my jets. Collaboration is key, after all.